uploading recent changes

+++ students/routes.js
@@ -17,14 +17,21 @@ routes.get('/getstudents/', async (req, res) => {
 });
 
 routes.get('/getstudent/:id', async (req, res) => {
-    try {
-        student_id = req.params.id;
-        const student = await db.getStudent(student_id);
+    studentId = req.params.id;
 
-        res.json(student);
+    if (await db.studentIdExists(studentId)) {
 
-    } catch (error) {
-        res.json(error);
+        try {
+
+            const student = await db.getStudent(studentId);
+
+            res.json(student);
+
+        } catch (error) {
+            res.json(error);
+        }
+    } else {
+        res.json({ status: `Student with id ${studentId} does not exist.` });
     }
 });
 
@@ -35,7 +42,7 @@ routes.post('/addstudent/', async (req, res) => {
         const addedStudent = await db.addStudent(studentData);
 
         res.json({ 'Added student': addedStudent });
-        
+
     } catch (error) {
         res.json(error);
     }
@@ -47,26 +54,35 @@ routes.put('/putstudent/:id', async (req, res) => {
     const age = req.body.age;
     const studentId = req.params.id;
 
-    try {
-        const updatedStudent = await db.updateStudent(firstname, lastname, age, studentId);
+    if (await db.studentIdExists(studentId)) {
 
-        res.json({ 'Old student information': updatedStudent[0], 'New student information': updatedStudent[1] });
+        try {
+            const updatedStudent = await db.updateStudent(firstname, lastname, age, studentId);
 
-    } catch (error) {
-        res.json(error)
+            res.json({ 'Old student information': updatedStudent[0], 'New student information': updatedStudent[1] });
+
+        } catch (error) {
+            res.json(error)
+        }
+    } else {
+        res.json({ status: `Student with id ${studentId} does not exist.` });
     }
 });
 
 routes.delete('/delstudent/:id', async (req, res) => {
     const studentId = req.params.id;
 
-    try {
-        const deletedStudent = await db.deleteStudent(studentId);
+    if (await db.studentIdExists(studentId)) {
+        try {
+            const deletedStudent = await db.deleteStudent(studentId);
 
-        res.json({ 'Deleted student': deletedStudent });
+            res.json({ 'Deleted student': deletedStudent });
 
-    } catch (error) {
-        res.json(error);
+        } catch (error) {
+            res.json(error);
+        }
+    } else {
+        res.json({ status: `Student with id ${studentId} does not exist.` });
     }
 });
